requirements.txt
nicegui
openai==0.28
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
pyjwt
